ne znam zasto authentification manager ne radi

+++ PKIService/bezbednost-backend/src/main/java/com/example/bezbednostbackend/service/RefreshTokenService.java
@@ -6,7 +6,7 @@ import java.util.Optional;
 
 public interface RefreshTokenService {
     Optional<RefreshToken> findByToken(String token);
-    RefreshToken createRefreshToken(Long userId);
+    RefreshToken createRefreshToken(Integer userId);
     RefreshToken verifyExpiration(RefreshToken token);
-    int deleteByUserId(Long userId);
+    int deleteByUserId(Integer userId);
 }
